+++ src/Exceptions/DgiiRepositoryException.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace PlatinumPlace\LaravelDgii\Exceptions;
+
+use RuntimeException;
+
+class DgiiRepositoryException extends RuntimeException
+{
+    //
+}